+++ libs/libcore/src/column.cpp
@@ -191,7 +191,7 @@ void Column::setSequence(BaseObject *seq)
 							.arg(this->getTypeName())
 							.arg(BaseObject::getTypeName(ObjectType::Sequence)),
 							ErrorCode::AsgInvalidObjectType,__PRETTY_FUNCTION__,__FILE__,__LINE__);
-		else if(!type.isIntegerType())
+		else if(!type.isIntegerType() && !type.isNumericType())
 			throw Exception(Exception::getErrorMessage(ErrorCode::IncompColumnTypeForSequence)
 							.arg(seq->getName(true))
 							.arg(this->obj_name),
@@ -314,8 +314,13 @@ QString Column::getAlterCode(BaseObject *object)
 
 		if(!this->type.isEquivalentTo(col->type) ||
 				(this->type.isEquivalentTo(col->type) &&
-				 ((this->type.hasVariableLength() && (this->type.getLength()!=col->type.getLength())) ||
-					(this->type.acceptsPrecision() && (this->type.getPrecision()!=col->type.getPrecision())))))
+
+				 ((this->type.hasVariableLength() &&
+					 this->type.getLength() != col->type.getLength()) ||
+
+					(this->type.acceptsPrecision() &&
+					 col->type.getPrecision() > 0 &&
+					 this->type.getPrecision() != col->type.getPrecision()))))
 			attribs[Attributes::Type]=col->type.getSourceCode(SchemaParser::SqlCode);
 
 		if(col->sequence)