+++ libs/libgui/src/tools/databaseimporthelper.cpp
@@ -570,6 +570,9 @@ void DatabaseImportHelper::importDatabase()
 		dbmodel->setLoadingModel(true);
 		dbmodel->setObjectListsCapacity(creation_order.size());
 
+		cached_names.clear();
+		cached_signatures.clear();
+
 		retrieveSystemObjects();
 		retrieveUserObjects();
 		createObjects();
@@ -850,7 +853,8 @@ QString DatabaseImportHelper::getDependencyObject(const QString &oid, ObjectType
 			if(use_signature)
 				obj_name = obj_attr[Attributes::Signature] = getObjectName(oid, true);
 			else
-				obj_name = obj_attr[Attributes::Name] = getObjectName(oid);
+				//obj_name = obj_attr[Attributes::Name] = getObjectName(oid);
+				obj_name = getObjectName(oid);
 
 			/* If the attributes of the dependency exists but it was not created on the model yet,
 				 pgModeler will create it and it's dependencies recursively */
@@ -2673,8 +2677,7 @@ void DatabaseImportHelper::createColumns(attribs_map &attribs, std::vector<unsig
 	//Retrieving columns if they were not retrieved yet
 	if(columns[attribs[Attributes::Oid].toUInt()].empty() && auto_resolve_deps)
 	{
-		QString sch_name;
-		sch_name = getDependencyObject(attribs[Attributes::SchemaOid], ObjectType::Schema, true, auto_resolve_deps, false);
+		QString sch_name = getDependencyObject(attribs[Attributes::SchemaOid], ObjectType::Schema, true, auto_resolve_deps, false);
 		retrieveTableColumns(sch_name, attribs[Attributes::Name]);
 	}
 
@@ -2834,8 +2837,8 @@ void DatabaseImportHelper::assignSequencesToColumns()
 			col=dynamic_cast<Column *>(tab_obj);
 
 			//Translating the default value nextval('sequence'::regclass)
-			if(col->getType().isIntegerType() &&
-					col->getDefaultValue().contains(QString("nextval(")))
+			if((col->getType().isIntegerType() || col->getType().isNumericType()) &&
+					col->getDefaultValue().contains("nextval("))
 			{
 				QString seq_name=col->getDefaultValue();
 				Sequence *seq=nullptr;
@@ -2990,7 +2993,6 @@ QString DatabaseImportHelper::getObjectName(const QString &oid, bool signature_f
 {
 	unsigned obj_oid=oid.toUInt();
 	attribs_map obj_attr = getObjectAttributes(obj_oid);
-	static std::map<unsigned, QString> cached_names, cached_signatures;
 
 	if(obj_oid == 0 || obj_attr.empty())
 		return "";